Add AbstractPluginCommand
