+++ src/Rocketeer/Console/Commands/Development/TinkerCommand.php
@@ -11,7 +11,7 @@
 namespace Rocketeer\Console\Commands\Development;
 
 use Psy\Shell;
-use Rocketeer\Abstracts\AbstractCommand;
+use Rocketeer\Abstracts\Commands\AbstractCommand;
 use Rocketeer\Console\TinkerApplication;
 
 class TinkerCommand extends AbstractCommand
