+++ src/Rocketeer/Console/Commands/Plugins/InstallCommand.php
@@ -10,11 +10,16 @@
 
 namespace Rocketeer\Console\Commands\Plugins;
 
-use Rocketeer\Abstracts\AbstractCommand;
+use Rocketeer\Abstracts\Commands\AbstractPluginCommand;
 use Symfony\Component\Console\Input\InputArgument;
 
-class InstallCommand extends AbstractCommand
+class InstallCommand extends AbstractPluginCommand
 {
+    /**
+     * @type string
+     */
+    protected $pluginTask = 'Installer';
+
     /**
      * The default name
      *
@@ -28,34 +33,4 @@ class InstallCommand extends AbstractCommand
      * @type string
      */
     protected $description = 'Install a plugin';
-
-    /**
-     * Whether the command's task should be built
-     * into a pipeline or run straight
-     *
-     * @type boolean
-     */
-    protected $straight = true;
-
-    /**
-     * Run the tasks
-     *
-     * @return integer
-     */
-    public function fire()
-    {
-        return $this->fireTasksQueue('Plugins\Installer');
-    }
-
-    /**
-     * Get the console command arguments.
-     *
-     * @return string[][]
-     */
-    protected function getArguments()
-    {
-        return array(
-            ['package', InputArgument::REQUIRED, 'The package to publish the configuration for'],
-        );
-    }
 }
