+++ src/Rocketeer/Console/Commands/Plugins/PublishCommand.php
@@ -10,7 +10,7 @@
 
 namespace Rocketeer\Console\Commands\Plugins;
 
-use Rocketeer\Abstracts\AbstractCommand;
+use Rocketeer\Abstracts\Commands\AbstractCommand;
 use Rocketeer\Services\Ignition\Plugins;
 use Symfony\Component\Console\Input\InputArgument;
