+++ src/Rocketeer/Console/Commands/StrategiesCommand.php
@@ -10,7 +10,7 @@
 
 namespace Rocketeer\Console\Commands;
 
-use Rocketeer\Abstracts\AbstractCommand;
+use Rocketeer\Abstracts\Commands\AbstractCommand;
 
 /**
  * Lists the available options for each strategy
